+++ pkg/lib/src/syscall.rs
@@ -130,6 +130,6 @@ pub fn sys_brk(addr: Option<usize>) -> Option<usize> {
     const BRK_FAILED: usize = !0;
     match syscall!(Syscall::Brk, addr.unwrap_or(0)) {
         BRK_FAILED => None,
-        ret => Some(ret as usize),
+        ret => Some(ret),
     }
 }
